[refactoring] change folder tree
