+++ SSD_TestShell/command/command.h
@@ -4,7 +4,7 @@
 #include <string>
 #include "ssd_handler.h"
 #include "utils.h"
-#include "logger/logger.h"
+#include "logger.h"
 
 #define interface class
