+++ SSD_TestShell/command/erase_command.cpp
