+++ SSD_TestShell/command/full_write_and_read_compare_command.cpp
