+++ SSD_TestShell/test/runner_test.cpp
@@ -1,9 +1,10 @@
 #include "gmock/gmock.h"
-#include "../runner.h"
-#include "../full_write_and_read_compare_command.h"
-#include "../partial_lba_write_command.h"
-#include "../write_read_aging_command.h"
-#include "../erase_and_write_aging_command.h"
+#include "runner.h"
+#include "full_write_and_read_compare_command.h"
+#include "partial_lba_write_command.h"
+#include "write_read_aging_command.h"
+#include "erase_and_write_aging_command.h"
+#include "ssd_handler.h"
 
 #include <iostream>
 #include <sstream>
