+++ SSD_TestShell/util/utils.cpp
